KOCO-bench-en/domain_code_generation/scripts/agent/dataset/dataset_remove_implementation_remove_test_en/verl/test_examples/PURE/code/requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
flash-attn
hydra-core
liger-kernel
numpy
pandas
peft
pyarrow>=15.0.0
pybind11
pylatexenc
ray[data,train,tune,serve]
tensordict<0.10
torchdata
transformers
vllm<=0.9.2
wandb
